Adding "Export SVG" button
This is a very basic implementation of SVG export.

It looks into DOM elements to figure out the positions and dimensions,
and then re-creates the tables as plain SVG <rect> and <text>.

Works fine on both Firefox and Chrome.

Known issues:

* The positions are incorrect if the browser has some zoom applied. In
fact, the display is already incorrect in the browser itself (the SVG
paths are not correctly aligned with the elements).
* Rounded corners are not supported. Implementing that (properly) would
require SVG clipping or some extra tricks. Not worth the extra trouble
for now.
* The borders are 2-pixel wide because otherwise they would not align to
the pixel grid (1-pixel border would cover two half pixels).
* There are no translations for other languages.

Related issues:

* https://github.com/ondras/wwwsqldesigner/issues/131
* https://github.com/ondras/wwwsqldesigner/issues/132